Fix media issue
Haven uploads media to `/haven/data/blossom` **not** `/haven/blossom`. As a result the binding was just binding to empty directories and thus media uploads were not persistent. You can verify this by checking an existing deployments data directory on the host machine and seeing if any of your uploads appear in the blossom directory - they won't. The new binding ensures persistence.